requirements.txt
playwright==1.49.0
psycopg2-binary>=2.9.0
requests>=2.28.0
